fixed label issue and added counter for each class
When yolo.py is configured to detect more than one class, an error
occurs where every object is labeled the same class. This is fixed by
modifying detection.py and detection_yolo.py in the deep_sort folder
so that the correct class can be retrieved for each detection.

Also added a counter for each class that will dynamically appear below
the total objects counter