fixed FPS counter, added text scaling and cam ID
Fixed the flashing FPS on the YOLO stream due to the camera stream frames being
occasionally read at the same time as writing the FPS to the frame.
Camera stream FPS is now written as they are received by the server thread to
fix this issue. This also fixed an issue where the camera stream FPS is
not actually showing the FPS of the incoming stream from the camera
client.

All text displayed now also scales in spacing and size depending on the resolution
of the incoming camera stream, making it more readable without having
to manually change everything.

Camera ID for each stream is now displayed depending on what is used as
the cam_id for each camera client. This makes it easier to see which stream is
which, since the device numbers cannot be changed given
how ports are assigned to each client.